fix: improve attachment_size validation and fix compilation warnings
- Add explicit negative value validation for attachment_size before casting to size_t
- Remove redundant GetAttachmentSize calls in ProcessRpcRequest
- Mark GetAttachmentSizeFromDump as unused to avoid compilation warnings
- Improve validation logic in rpc_replay to validate before casting
- Revert nshead_meta.proto attachment_size to int32 as per maintainer feedback

Fixes issue #1348

+++ tools/rpc_replay/rpc_replay.cpp
@@ -189,8 +189,10 @@ static void* replay_thread(void* arg) {
                 } else if (sample->meta.has_attachment_size()) {
                     attachment_size = static_cast<int64_t>(sample->meta.attachment_size());
                 }
+                // Validate attachment_size: check for negative values and size overflow
+                // Explicitly validate the range before casting to size_t
                 if (attachment_size > 0 && 
-                    static_cast<size_t>(attachment_size) < sample->request.size()) {
+                    attachment_size < static_cast<int64_t>(sample->request.size())) {
                     sample->request.cutn(
                         &req.serialized_data(),
                         sample->request.size() - static_cast<size_t>(attachment_size));
